Refactored most MultiMC code to LauncherUtils
