+++ src/main/kotlin/link/infra/packwiz/installer/UpdateManager.kt
@@ -2,7 +2,6 @@ package link.infra.packwiz.installer
 
 import com.google.gson.GsonBuilder
 import com.google.gson.JsonIOException
-import com.google.gson.JsonParser
 import com.google.gson.JsonSyntaxException
 import com.moandjiezana.toml.Toml
 import link.infra.packwiz.installer.DownloadTask.Companion.createTasksFromIndex
@@ -61,7 +60,7 @@ class UpdateManager internal constructor(private val opts: Options, val ui: IUse
 		checkOptions()
 
 		ui.submitProgress(InstallProgress("Loading manifest file..."))
-		val gson = GsonBuilder().registerTypeAdapter(Hash::class.java, Hash.TypeHandler()).setPrettyPrinting().create()
+		val gson = GsonBuilder().registerTypeAdapter(Hash::class.java, Hash.TypeHandler()).create()
 		val manifest = try {
 			gson.fromJson(FileReader(Paths.get(opts.packFolder, opts.manifestFile).toString()),
 					ManifestFile::class.java)
@@ -99,94 +98,24 @@ class UpdateManager internal constructor(private val opts: Options, val ui: IUse
 			handleCancellation()
 		}
 
+		// Launcher checks
+		val lu = LauncherUtils(opts, ui)
+
 		// MultiMC MC and loader version checker
-		val multimcManifestPath = Paths.get(opts.multimcFolder, "mmc-pack.json").toString()
-		val multimcManifestFile = File(multimcManifestPath)
-
-		if (multimcManifestFile.exists()) {
-			ui.submitProgress(InstallProgress("Loading MultiMC pack file..."))
-
-			val multimcManifest = try {
-				JsonParser.parseReader(multimcManifestFile.reader())
-			} catch (e: JsonIOException) {
-				ui.showErrorAndExit("Cannot read the MultiMC pack file", e)
-			} catch (e: JsonSyntaxException) {
-				ui.showErrorAndExit("Invalid MultiMC pack file", e)
-			}.asJsonObject
-
-			Log.info("Loaded MultiMC config")
-
-			// We only support format 1, if it gets updated in the future we'll have to handle that
-			// There's only version 1 for now tho, so that's good
-			if (multimcManifest["formatVersion"].asInt != 1) {
-				ui.showErrorAndExit("Invalid MultiMC format version")
+		ui.submitProgress(InstallProgress("Loading MultiMC pack file..."))
+		try {
+			when (lu.handleMultiMC(pf, gson)) {
+				LauncherUtils.LauncherStatus.Cancelled -> cancelled = true
+				LauncherUtils.LauncherStatus.NotFound -> Log.info("MultiMC not detected")
 			}
+			handleCancellation()
+		} catch (e: Exception) {
+			ui.showErrorAndExit(e.message!!, e)
+		}
 
-			var manifestModified = false
-			val modLoaders = hashMapOf("net.minecraft" to "minecraft", "net.minecraftforge" to "forge", "net.fabricmc.fabric-loader" to "fabric", "org.quiltmc.quilt-loader" to "quilt", "com.mumfrey.liteloader" to "liteloader")
-			val modLoadersClasses = modLoaders.entries.associate{(k,v)-> v to k}
-			var modLoaderFound = false
-			val modLoadersFound = HashMap<String, String>() // Key: modLoader, Value: Version
-			val components = multimcManifest["components"].asJsonArray
-			for (componentObj in components) {
-				val component = componentObj.asJsonObject
-
-				val version = component["version"].asString
-				// If we find any of the modloaders we support, we save it and check the version
-				if (modLoaders.containsKey(component["uid"].asString)) {
-					val modLoader = modLoaders.getValue(component["uid"].asString)
-					if (modLoader != "minecraft")
-						modLoaderFound = true // Only set to true if modLoader isn't Minecraft
-					modLoadersFound[modLoader] = version
-					if (version != pf.versions?.get(modLoader)) {
-						manifestModified = true
-						component.addProperty("version", pf.versions?.get(modLoader))
-					}
-				}
-			}
-
-			// If we can't find the mod loader in the MultiMC file, we add it
-			if (!modLoaderFound) {
-				// Using this filter and loop to handle multiple handlers
-				for ((_, loader) in modLoaders
-						.filter { it.value != "minecraft" && !modLoadersFound.containsKey(it.value) && pf.versions?.containsKey(it.value) == true }
-				) {
-					components.add(gson.toJsonTree(hashMapOf("uid" to modLoadersClasses.get(loader), "version" to pf.versions?.get(loader))))
-				}
-			}
-
-			// If mc version change detected, and fabric mappings are found, delete them, MultiMC will add and re-dl the correct one
-			if (modLoadersFound["minecraft"] != pf.versions?.getValue("minecraft"))
-				components.find { it.asJsonObject["uid"].asString == "net.fabricmc.intermediary" }?.asJsonObject?.let { components.remove(it) }
-
-			if (manifestModified) {
-				// The manifest has been modified, so before saving it we'll ask the user
-				// if they wanna update it, continue without updating it, or exit
-				val oldVers = modLoadersFound.map { Pair(it.key, it.value) }
-				val newVers = pf.versions!!.map { Pair(it.key, it.value) }
-
-
-				when (ui.showUpdateConfirmationDialog(oldVers, newVers)) {
-					IUserInterface.UpdateConfirmationResult.CANCELLED -> {
-						cancelled = true
-					}
-					IUserInterface.UpdateConfirmationResult.CONTINUE -> {
-						cancelledStartGame = true
-					}
-					else -> {} // Compiler is giving warning about "non-exhaustive when", so i'll just add an empty one
-				}
-				handleCancellation()
-
-				multimcManifestFile.writeText(gson.toJson(multimcManifest))
-				Log.info("Updated modpack Minecrafts and/or the modloaders version")
-			}
-
-			if (ui.cancelButtonPressed) {
-				showCancellationDialog()
-				handleCancellation()
-			}
-		} else {
-			Log.info("MultiMC installation not detected... Tried looking for it in $multimcManifestPath")
+		if (ui.cancelButtonPressed) {
+			showCancellationDialog()
+			handleCancellation()
 		}
 
 		ui.submitProgress(InstallProgress("Checking local files..."))
